removes old implementation
