--- src/lib/qwik-trpc.ts
@@ -1,138 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/**
- * @see https://trpc.io/blog/tinyrpc-client
- */
-import {
-  type Action,
-  type RequestEventCommon,
-  type RequestEventLoader,
-} from "@builder.io/qwik-city";
-import type {
-  AnyMutationProcedure,
-  AnyProcedure,
-  AnyQueryProcedure,
-  AnyRouter,
-  inferProcedureInput,
-  inferProcedureOutput,
-  ProcedureRouterRecord,
-  TRPCError,
-} from "@trpc/server";
-import type { ZodIssue } from "zod";
-import { getTrpcFromEvent } from "~/server/loaders";
-import type { TypedServerFunction } from "~/utils/types";
-
-type ProxyCallbackOptions = {
-  path: string[];
-  args: unknown[];
-};
-
-type ProxyCallback = (opts: ProxyCallbackOptions) => unknown;
-
-type TrpcProcedureOutput<TProcedure extends AnyProcedure> =
-  | {
-      result: inferProcedureOutput<TProcedure>;
-      status: "success";
-    }
-  | {
-      code: ZodIssue["code"];
-      status: "error";
-      issues: ZodIssue[];
-    };
-
-type DecorateProcedure<TProcedure extends AnyProcedure> =
-  TProcedure extends AnyQueryProcedure
-    ? {
-        loader: (
-          event: RequestEventLoader,
-          input: inferProcedureInput<TProcedure>
-        ) => TrpcProcedureOutput<TProcedure>;
-        query: () => TypedServerFunction<
-          inferProcedureInput<TProcedure>,
-          TrpcProcedureOutput<TProcedure>
-        >;
-      }
-    : TProcedure extends AnyMutationProcedure
-    ? {
-        action$: () => Action<
-          TrpcProcedureOutput<TProcedure>,
-          inferProcedureInput<TProcedure>,
-          false
-        >;
-        mutate: () => TypedServerFunction<
-          inferProcedureInput<TProcedure>,
-          TrpcProcedureOutput<TProcedure>
-        >;
-      }
-    : never;
-
-type DecoratedProcedureRecord<TProcedures extends ProcedureRouterRecord> = {
-  [TKey in keyof TProcedures]: TProcedures[TKey] extends AnyRouter
-    ? DecoratedProcedureRecord<TProcedures[TKey]["_def"]["record"]>
-    : TProcedures[TKey] extends AnyProcedure
-    ? DecorateProcedure<TProcedures[TKey]>
-    : never;
-};
-
-type HandleRequestArgs = {
-  event: RequestEventCommon;
-  args: any;
-  dotPath: string[];
-};
-
-export const createTrpcServerApi = <TRouter extends AnyRouter>() => {
-  const createRecursiveProxy = (callback: ProxyCallback, path: string[]) => {
-    const proxy: unknown = new Proxy(() => void 0, {
-      apply(_1, _2, args) {
-        return callback({ args, path });
-      },
-      get(_obj, key) {
-        if (typeof key !== "string") {
-          return undefined;
-        }
-        return createRecursiveProxy(callback, [...path, key]);
-      },
-    });
-
-    return proxy;
-  };
-
-  const handleRequest = async ({ event, args, dotPath }: HandleRequestArgs) => {
-    const trpc = await getTrpcFromEvent(event);
-
-    const fnc = dotPath.reduce((prev, curr) => prev[curr], trpc as any);
-
-    const safeParse = (data: string) => {
-      try {
-        return JSON.parse(data);
-      } catch {
-        return data;
-      }
-    };
-
-    try {
-      const result = await fnc(args);
-
-      return { result, status: "success" };
-    } catch (err) {
-      const trpcError = err as TRPCError;
-      const error = {
-        code: trpcError.code,
-        issues: safeParse(trpcError.message),
-        status: "error",
-      };
-      return error;
-    }
-  };
-
-  return createRecursiveProxy((opts) => {
-    const dotPath = opts.path.slice(0, -1);
-    const action = opts.path[opts.path.length - 1];
-
-    if (action === "loader") {
-      const event = opts.args[0] as RequestEventLoader;
-      const args = opts.args[1];
-
-      return handleRequest({ args, dotPath, event });
-    }
-  }, []) as DecoratedProcedureRecord<TRouter["_def"]["record"]>;
-};
