--- src/server/trpc/api.ts
@@ -1,4 +1,0 @@
-import { createTrpcServerApi } from "~/lib/qwik-trpc";
-import type { AppRouter } from "./router";
-
-export const trpc = createTrpcServerApi<AppRouter>();
